Create Project_Flying_Fish.pde
This is a java project done using processing.

The user can draw shapes onto the screen, the shapes are drawn to the location of the mouse.

In this project, the user can press six numbers that all do six different drawing actions. The number keys range from the numbers 1 to 6.